Rename EnsureDeviceAppInteraction -> EnsureDeviceApp, and get rid of the old one
